Animation & Color
The first major difference is that we changed some of the values in our model
from `Float` to `Animation`. This is because we can always fetch its "current"
value using `getTo` and update with `retarget`.

Another feature is that the background color is an HSL value. The hue matches
the user's selected team (use `t` to cycle). The saturation corresponds to
latitude and the light is based on the longitude and time of day.
Notice that we're using nifty trigonometry to calculate some of these values
based on `persp.loc`.